Fix GCL 64-bit pointer truncation: change value_p to intptr_t (Option 2)
The fundamental issue: GCL_GetNextValue stored script data pointers in
int *value_p (4 bytes), truncating 64-bit pointers. The circular ptr table
workaround was fragile — entries got overwritten before bind triggers fired.

Fix: change GCL_GetNextValue's value_p parameter from int* to intptr_t*.
Pointers into GCL script data (GCLCODE_STRING, GCLCODE_SCRIPT_DATA,
GCLCODE_PARAMETER) are now stored directly as full 64-bit values.

Updated all callers in libgcl_fix/ (basic.c, command.c, expr.c, variable.c,
parse.c) and game_script_fix.c to use intptr_t for value variables.

The gcl_ptr_table is kept for backward compatibility but pointer-type
GCL values no longer need it — they go directly into the intptr_t variable.

This fixes:
- GCL_Command_if block pointer truncation (crash on stage scripts)
- HZD bind GCL block execution (camera zone transitions)
- All GCL script data references that contain embedded pointers

+++ port/libgcl_fix/command.c
@@ -113,7 +113,7 @@ static int GCL_Proc(unsigned char *ptr)
     long     argbuf[8];
     GCL_ARGS args;
     int      code;
-    int      value;
+    intptr_t value;
     int      arg_idx;
 
     int b1 = ptr[0];
